% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/db_functions.R
\name{dbHeadTable}
\alias{dbHeadTable}
\title{Head}
\usage{
dbHeadTable(path, dbName, table, n = 6)
}
\arguments{
\item{path}{path to database}

\item{dbName}{database name}

\item{table}{table name}

\item{n}{number of lines}
}
\value{
head of table
}
\description{
Head of the database table, similar to head function.
}
\details{
Details follow.
}
\author{
Maarten van Iterson
}

