% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/db_functions.R
\name{dbInfo}
\alias{dbInfo}
\title{Info}
\usage{
dbInfo(path, dbName)
}
\arguments{
\item{path}{path to database}

\item{dbName}{database name}
}
\value{
info
}
\description{
Gives some info of tables and fields of the database.
}
\details{
Details follow.
}
\author{
Maarten van Iterson
}

