% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/download.R
\name{downloadTargets}
\alias{downloadTargets}
\title{Download predicted microRNA targets}
\usage{
downloadTargets(destDir)
}
\arguments{
\item{destDir}{directory where you want to store the predicted microRNA target files}
}
\description{
Download predicted microRNA targets
original NAR paper resources:
http://genie.weizmann.ac.il/pubs/mir07/catalogs/PITA_targets_mm9_0_0_TOP.tab.gz
ftp://ftp.ebi.ac.uk/pub/databases/microcosm/v5/arch.v5.gff.mus_musculus.zip
http://www.targetscan.org//mmu_60//mmu_60_data_download/Conserved_Site_Context_Scores.txt.zip
updated resources:
http://www.targetscan.org/vert_61/vert_61_data_download/Conserved_Site_Context_Scores.txt.zip
}
\examples{
\dontrun{
##requires write access to the installed package directory
##if you don't have define your own directory
library(miRNAmRNA)
dir.create(file.path(path.package("miRNAmRNA"),"extdata"))
downloadTargets(file.path(path.package("miRNAmRNA"),"extdata"))}
}
\author{
Maarten van Iterson
}

