\name{gtTable}
\alias{gtTable}
\title{Extract gtTable}
\usage{
  gtTable(object)
}
\arguments{
  \item{object}{gt-object}
}
\value{
  gt-table
}
\description{
  gtTable extracts all the targets and their p-values from
  a GT object.
}
\details{
  Details follow.
}
\author{
  Maarten van Iterson, Sander Bervoets
}

