% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/db_functions.R
\name{overlapMircoRNAs}
\alias{overlapMircoRNAs}
\title{Find unique microRNAs old}
\usage{
overlapMircoRNAs(path, dbName, tables = NULL, target, numOverlapping = 2,
  full = TRUE)
}
\arguments{
\item{path}{to database}

\item{dbName}{database name}

\item{tables}{tables in the database}

\item{target}{specific target}

\item{full}{return whole matrix}
}
\value{
unique microRNAs
}
\description{
Find unique microRNAs old
}
\details{
Details will follow
}
\author{
Maarten van Iterson
}

