\name{uniqueMirs}
\alias{uniqueMirs}
\title{Unique mirs}
\usage{
  uniqueMirs(path, dbName)
}
\arguments{
  \item{path}{path to database}

  \item{dbName}{database name}
}
\value{
  vector of all unique mirs
}
\description{
  function to get all unique mirs from in the database.
}
\details{
  Details follow.
}
\author{
  Maarten van Iterson, Sander Bervoets
}

