\name{uniqueMirs_new}
\alias{uniqueMirs_new}
\title{Find unique microRNAs}
\usage{
  uniqueMirs_new(path, dbName, target, tables = NULL)
}
\arguments{
  \item{path}{to database}

  \item{dbName}{database name}

  \item{target}{specific target}

  \item{tables}{tables in the database}
}
\value{
  unique microRNAs
}
\description{
  Find unique microRNAs
}
\details{
  Details will follow
}
\author{
  Maarten van Iterson
}

