\name{uniqueTargets}
\alias{uniqueTargets}
\title{Unique mirs}
\usage{
  uniqueTargets(path, dbName, mir, tables = NULL)
}
\arguments{
  \item{path}{path to database}

  \item{dbName}{database name}

  \item{mir}{microRNA for which unique targets will be
  obtained}

  \item{tables}{NULL or one or more table names in the
  database}
}
\value{
  vector of unique targets
}
\description{
  function to get all unique targets for a specific mir.
}
\details{
  Details follow.
}
\author{
  Maarten van Iterson, Sander Bervoets
}

