\name{dbHeadTable}
\alias{dbHeadTable}
\title{Head}
\usage{
  dbHeadTable(path, dbName, table, n = 6)
}
\arguments{
  \item{dbName}{database name}

  \item{table}{table name}

  \item{path}{path to database}

  \item{n}{number of lines}
}
\value{
  head of table
}
\description{
  Head of the database table, similar to head function.
}
\details{
  Details follow.
}
\author{
  Maarten van Iterson
}

