\name{dbInfo}
\alias{dbInfo}
\title{Info}
\usage{
  dbInfo(path, dbName)
}
\arguments{
  \item{dbName}{database name}

  \item{path}{path to database}
}
\value{
  info
}
\description{
  Gives some info of tables and fields of the database.
}
\details{
  Details follow.
}
\author{
  Maarten van Iterson
}

