% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mirnamrna.R
\name{gtTable}
\alias{gtTable}
\title{Extract gtTable}
\usage{
gtTable(object)
}
\arguments{
\item{object}{gt-object}
}
\value{
gt-table
}
\description{
gtTable extracts all the targets and their p-values from a GT object.
}
\details{
Details follow.
}
\author{
Maarten van Iterson, Sander Bervoets
}

