\name{overlapMircoRNAs}
\alias{overlapMircoRNAs}
\title{Find unique microRNAs old}
\usage{
  overlapMircoRNAs(path, dbName, tables = NULL, target,
    numOverlapping = 2, full = TRUE)
}
\arguments{
  \item{path}{to database}

  \item{dbName}{database name}

  \item{target}{specific target}

  \item{tables}{tables in the database}

  \item{full}{return whole matrix}
}
\value{
  unique microRNAs
}
\description{
  Find unique microRNAs old
}
\details{
  Details will follow
}
\author{
  Maarten van Iterson
}

