% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mirnamrna.R
\name{runrgt}
\alias{runrgt}
\title{Run the global test on the list of mirs}
\usage{
runrgt(targets, Y, X, A)
}
\arguments{
\item{targets}{list of targets}

\item{Y}{miRNA expression}

\item{X}{mRNA expression}

\item{A}{data.frame containing predicted microRNA target pairs}
}
\value{
list of microRNAs and targets
}
\description{
reversed version rungt is a wrapper for the gt() function
}
\details{
Details follow.
}
\author{
Maarten van Iterson
}

